pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Tomi <tomascufaro@gmail.com>"]
readme = "README.md"
packages = [{include = "src"},]


[tool.poetry.dependencies]
python = ">=3.9,<3.11"
requests = "^2.30.0"
pandas = "1.5.3"
fire = "^0.5.0"
pyarrow = "^12.0.0"
ta = "^0.10.2"
notebook = "^6.5.4"
scikit-learn = "^1.2.2"
xgboost = "^1.7.5"
lightgbm = "^3.3.5"
optuna = "^3.1.1"
comet-ml = "3.32.6"
cerebrium = "^1.1.1"
torch = "1.13.1"
pydantic = "^2.1.1"
alpaca-py = "*"
mlflow = "*"
python-dotenv = "*"
ipykernel = "^6.29.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"